[ENG-1963] Update UI for CashApp (#180)

+++ LiquidityFinancial/Modules/LFDomain/Sources/RewardDomain/Entities/UserDonationSummaryEntity.swift
@@ -2,7 +2,7 @@ import Foundation
 
 public protocol UserDonationSummaryEntity {
   associatedtype APIDonationSticker: DonationStickerEntity
-  var donationAmount: Int? { get }
+  var donationAmount: Double? { get }
   var donationCount: Int? { get }
   var stickers: [APIDonationSticker] { get }
 }
